requirements.txt
# Google ADK Framework
google-adk>=1.19.0
google-genai>=1.45.0

# Core Dependencies
pydantic>=2.0.0
python-dotenv>=1.0.0

# Observability
structlog>=23.1.0

# Utilities
python-dateutil>=2.8.0
